Restore missing functionality to Container page (#244)
