+++ assets/src/components/cluster/containers/ContainerShell.tsx
@@ -1,206 +1,29 @@
-import {
-  Button,
-  CliIcon,
-  CloseIcon,
-  Input,
-  PencilIcon,
-  ToolIcon,
-  Tooltip,
-} from '@pluralsh/design-system'
-import {
-  useContext,
-  useEffect,
-  useRef,
-  useState,
-} from 'react'
-import { Flex, Form, Span } from 'honorable'
+import { Button, ToolIcon, Tooltip } from '@pluralsh/design-system'
+import { useContext, useRef, useState } from 'react'
+import { Div, Flex } from 'honorable'
 import { useParams } from 'react-router-dom'
 
-import styled from 'styled-components'
+import TerminalThemeSelector from 'components/terminal/TerminalThemeSelector'
 
 import { ShellContext, TerminalActions, TerminalScreen } from '../../terminal/Terminal'
 
+import { CODELINE_HEIGHT, ShellCommandEditor } from './ShellCommandEditor'
+
 export const DEFAULT_COMMAND = '/bin/sh'
 
-const CodeInput = styled(Input)(({ theme }) => ({
-  display: 'flex',
-  flexGrow: 1,
-  flexShrink: 1,
-  alignItems: 'center',
-  'input, div': {
-    ...theme.partials.text.code,
-  },
-  '& > div': {
-    maxWidth: '70%',
-    justifyContent: 'left',
-    overflow: 'hidden',
-    whiteSpace: 'nowrap',
-    '&, *': {
-      overflow: 'hidden',
-      textOverflow: 'ellipsis',
-      whiteSpace: 'nowrap',
-      textAlign: 'left',
-      justifyContent: 'left',
-    },
-  },
-}))
-
-const CODELINE_HEIGHT = 42
-
-const CodeLine = styled(({ children, ...props }) => (
-  <div {...props}>
-    <div>{children}</div>
-  </div>
-))(({ theme }) => ({
-  ...theme.partials.text.code,
-  position: 'relative',
-  overflowX: 'auto',
-  overflowY: 'hidden',
-  height: CODELINE_HEIGHT,
-  whiteSpace: 'nowrap',
-  border: theme.borders.input,
-  borderRadius: theme.borderRadiuses.medium,
-  color: theme.colors['text-xlight'],
-  backgroundColor: theme.colors['fill-one'],
-  div: {
-    display: 'flex',
-    alignItems: 'center',
-    position: 'absolute',
-    paddingLeft: theme.spacing.small,
-    paddingRight: theme.spacing.small,
-    top: 0,
-    height: CODELINE_HEIGHT - theme.borderWidths.default * 2,
-  },
-}))
-
-function truncRight(str: string, amt: number) {
-  return str.slice(str.length - amt, str.length)
-}
-function truncLeft(str: string, amt: number) {
-  return str.slice(0, amt)
-}
-
-const CodeWrap = styled.div<{ $isEditing: boolean }>(({ $isEditing }) => ({
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'stretch',
-  overflowX: 'hidden',
-  ...($isEditing ? { width: 0 } : { flexGrow: 1, flexShrink: 1 }),
-}))
-
-export function ShellCommand({
-  command,
-  setCommand,
-  isDefault,
-  defaultCommand,
-}: {
-  command: string
-  setCommand: (arg: string | null) => void
-  isDefault: boolean
-  defaultCommand: string
-}) {
-  const [isEditing, setIsEditing] = useState(false)
-  const [inputVal, setInputVal] = useState(command)
-  const { namespace, name, container } = useParams()
-  const inputWrapRef = useRef<HTMLDivElement>()
-
-  useEffect(() => {
-    if (!isEditing) {
-      setInputVal(command)
-    }
-  }, [command, isEditing])
-
-  useEffect(() => {
-    if (isEditing && inputWrapRef.current) {
-      const input = inputWrapRef.current.querySelector('input')
-
-      input?.focus()
-    }
-  }, [isEditing])
-
-  const commandLeft = `kubectl exec ${name} -it -n ${namespace} -c ${container} -- `
-  const commandLeftTrunc = (
-    <span>
-      {truncLeft(commandLeft, 13)}
-      <Span color="text-xlight">...</Span>
-      {truncRight(commandLeft, 20)}
-    </span>
-  )
-
+export function HeaderIconButton({ tooltipProps, ...props }) {
+  // Button needs div wrapper for tooltip to work
   return (
-    <Form
-      display="flex"
-      gap="xsmall"
-      onSubmit={e => {
-        e.preventDefault()
-        setCommand(inputVal)
-        setIsEditing(false)
-      }}
-      width="100%"
-    >
-      <Flex
-        ref={inputWrapRef as any}
-        width="50px"
-        flex="1 1"
-      >
-        <CodeWrap
-          $isEditing={isEditing}
-          className="codeWrap"
-        >
-          <CodeLine className="codeEdit">{`${commandLeft}${command}`}</CodeLine>
-        </CodeWrap>
-        {isEditing && (
-          <CodeInput
-            placeholder={defaultCommand}
-            value={inputVal}
-            prefix={commandLeftTrunc}
-            onChange={({ target: { value } }) => {
-              setInputVal(value)
-            }}
-            onKeyDown={(e: KeyboardEvent) => {
-              if (e.code.toLowerCase() === 'escape') {
-                setIsEditing(false)
-              }
-            }}
-          />
-        )}
-      </Flex>
-      {!isEditing && (
+    <Tooltip {...tooltipProps}>
+      <Div height="100%">
         <Button
           floating
-          size="medium"
-          startIcon={<PencilIcon />}
-          onClick={() => {
-            setIsEditing(true)
-          }}
-        >
-          Edit
-        </Button>
-      )}
-      {isEditing && (
-        <Button
-          floating
-          size="medium"
-          startIcon={<CliIcon />}
-          type="submit"
-        >
-          Run
-        </Button>
-      )}
-      {(isEditing || !isDefault) && (
-        <Button
-          secondary
-          size="medium"
-          startIcon={<CloseIcon />}
-          onClick={() => {
-            setCommand(null)
-            setIsEditing(false)
-          }}
-        >
-          Reset
-        </Button>
-      )}
-    </Form>
+          width={CODELINE_HEIGHT}
+          height="100%"
+          {...props}
+        />
+      </Div>
+    </Tooltip>
   )
 }
 
@@ -234,28 +57,22 @@ function ShellWithContext() {
       height="100%"
       gap="medium"
     >
-      <Flex
-        gap="medium"
-      >
-        <ShellCommand
+      <Flex gap="medium">
+        <ShellCommandEditor
           command={command}
           defaultCommand={defaultCommand}
           isDefault={isDefault}
           setCommand={setCommand}
         />
-        <Tooltip label="Repair viewport">
-          <Button
-            floating
-            small
-            onClick={() => {
-              shellContext?.current?.handleResetSize()
-            }}
-          ><ToolIcon size={16} />
-          </Button>
-        </Tooltip>
-        {/* <TerminalThemeSelector /> */}
-      </Flex>
+        <HeaderIconButton
+          tooltipProps={{ label: 'Repair viewport' }}
+          onClick={shellContext.current?.handleResetSize()}
+        >
+          <ToolIcon size={16} />
+        </HeaderIconButton>
 
+        <TerminalThemeSelector />
+      </Flex>
       <TerminalScreen
         room={`pod:${namespace}:${name}:${container}`}
         command={command}
