+++ assets/src/components/terminal/TerminalThemeSelector.tsx
@@ -3,14 +3,14 @@ import { Flex } from 'honorable'
 import Fuse from 'fuse.js'
 
 import {
-  Button,
-  DropdownArrowIcon,
   Input,
   ListBoxItem,
   MagnifyingGlassIcon,
   Select,
   SprayIcon,
 } from '@pluralsh/design-system'
+
+import { HeaderIconButton } from 'components/cluster/containers/ContainerShell'
 
 import { normalizedThemes, themeNames } from './themes'
 import TerminalThemeContext from './TerminalThemeContext'
@@ -20,7 +20,7 @@ const fuse = new Fuse(themeNames, { threshold: 0.25 })
 function TerminalThemeSelector() {
   const [, setTerminalTheme] = useContext(TerminalThemeContext)
   const [search, setSearch] = useState('')
-  const [open, setOpen] = useState(false)
+  const [, setOpen] = useState(false)
   const results = fuse.search(search).map(x => x.item)
   const displayedThemes = results.length ? results : themeNames
 
@@ -32,28 +32,9 @@ function TerminalThemeSelector() {
       onSelectionChange={t => setTerminalTheme(t)}
       onOpenChange={o => setOpen(o)}
       triggerButton={(
-        <Button
-          tertiary
-          small
-          startIcon={<SprayIcon />}
-          endIcon={(
-            <DropdownArrowIcon
-              marginLeft="8px"
-              size={12}
-              style={open ? {
-                transform: 'rotate(180deg)',
-                transitionDuration: '.2s',
-                transitionProperty: 'transform',
-              } : {
-                transform: 'rotate(0)',
-                transitionDuration: '.2s',
-                transitionProperty: 'transform',
-              }}
-            />
-          )}
-        >
-          Change theme
-        </Button>
+        <HeaderIconButton tooltipProps={{ label: 'Change theme' }}>
+          <SprayIcon />
+        </HeaderIconButton>
       )}
       dropdownFooterFixed={(
         <Flex
